refactor(sql): remove gnrsql_ prefix from sub-package module names
Rename all modules in the gnrsql/ sub-package to drop the redundant
gnrsql_ prefix (e.g. gnrsql_helpers.py → helpers.py).
Update all intra-package imports, docstrings and README accordingly.

Ref #497
